chore: Update dependencies and restructure project files
This commit updates several development dependencies in package.json and package-lock.json to their latest versions, ensuring compatibility and security. Additionally, it modifies the output path in webpack.config.js to direct compiled files to the 'docs/js' directory instead of 'public/js'. The static file serving path in server.ts is also updated to reflect this change.

Furthermore, the footer in index.html is enhanced with a GitHub link for better visibility and accessibility.

No newline was added at the end of the file to maintain consistency with previous formatting.